+++ src/console/controllers/CacheController.php
@@ -22,27 +22,51 @@ class CacheController extends Controller
     // =========================================================================
 
     /**
-     * Clears the cache.
+     * Clears the cache (files only).
      *
      * @return int
      */
     public function actionClear(): int
     {
-        return $this->_clearFlushCache(false);
+        Blitz::$plugin->cache->emptyCache(false);
+
+        $this->stdout(Craft::t('blitz', 'Blitz cache successfully cleared.').PHP_EOL, Console::FG_GREEN);
+
+        return ExitCode::OK;
     }
 
     /**
-     * Flushes the cache.
+     * Flushes the cache (files and database records).
      *
      * @return int
      */
     public function actionFlush(): int
     {
-        return $this->_clearFlushCache(true);
+        Blitz::$plugin->cache->emptyCache(true);
+
+        $this->stdout(Craft::t('blitz', 'Blitz cache successfully flushed.').PHP_EOL, Console::FG_GREEN);
+
+        return ExitCode::OK;
     }
 
     /**
-     * Clears and warms the entire cache.
+     * Refreshes expired elements.
+     *
+     * @return int
+     */
+    public function actionRefreshExpired(): int
+    {
+        Blitz::$plugin->cache->invalidateCache();
+
+        Craft::$app->getQueue()->run();
+
+        $this->stdout(Craft::t('blitz', 'Expired Blitz cache successfully refreshed.').PHP_EOL, Console::FG_GREEN);
+
+        return ExitCode::OK;
+    }
+
+    /**
+     * Flushes and warms the entire cache.
      *
      * @return int
      */
@@ -62,12 +86,12 @@ class CacheController extends Controller
             return ExitCode::OK;
         }
 
+        // Get warm cache URLS
+        $urls = Blitz::$plugin->cache->getAllCacheableUrls();
+
         $this->stdout(Craft::t('blitz', 'Flushing Blitz cache.').PHP_EOL, Console::FG_GREEN);
 
         Blitz::$plugin->cache->emptyCache(true);
-
-        // Get warm cache URLS
-        $urls = Blitz::$plugin->cache->getAllCacheableUrls();
 
         $total = count($urls);
         $count = 0;
@@ -119,39 +143,6 @@ class CacheController extends Controller
         }
 
         $this->stdout(Craft::t('blitz', 'Blitz cache successfully warmed {success} files.', ['success' => $success]).PHP_EOL, Console::FG_GREEN);
-
-        return ExitCode::OK;
-    }
-
-    // Private Methods
-    // =========================================================================
-
-    /**
-     * Clears or flushes the cache.
-     *
-     * @param bool $flush
-     *
-     * @return int
-     */
-    public function _clearFlushCache(bool $flush = false): int
-    {
-        $settings = Blitz::$plugin->getSettings();
-
-        if (empty($settings->cacheFolderPath)) {
-            $this->stderr(Craft::t('blitz', 'Blitz cache folder path is not set.').PHP_EOL, Console::FG_RED);
-
-            return ExitCode::OK;
-        }
-
-        $this->stdout(Craft::t('blitz', '{action} Blitz cache.', [
-            'action' => $flush ? 'Flushing' : 'Clearing'
-        ]).PHP_EOL, Console::FG_GREEN);
-
-        Blitz::$plugin->cache->emptyCache($flush);
-
-        $this->stdout(Craft::t('blitz', 'Blitz cache successfully {action}.', [
-            'action' => $flush ? 'flushed' : 'cleared'
-        ]).PHP_EOL, Console::FG_GREEN);
 
         return ExitCode::OK;
     }