+++ src/utilities/CacheUtility.php
@@ -7,6 +7,8 @@ namespace putyourlightson\blitz\utilities;
 
 use Craft;
 use craft\base\Utility;
+use craft\helpers\FileHelper;
+use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\records\CacheRecord;
 
 class CacheUtility extends Utility
@@ -43,25 +45,37 @@ class CacheUtility extends Utility
      */
     public static function contentHtml(): string
     {
-        $options = [];
+        $sites = [];
+        $sitePaths = [];
 
-        $sites = Craft::$app->getSites()->getAllSites();
+        $allSites = Craft::$app->getSites()->getAllSites();
 
-        foreach ($sites as $site) {
-            $cacheRecordCount = CacheRecord::find()
-                ->where(['siteId' => $site->id])
-                ->count();
+        foreach ($allSites as $site) {
+            $path = Blitz::$plugin->file->getSitePath($site->id);
+            $count = is_dir($path) ? count(FileHelper::findFiles($path)) : 0;
 
-            $options[] = [
-                'label' => Craft::t('blitz', '{site} ({count} {n,plural,=1{page} other{pages}} cached)',
-                    ['site' => $site->name, 'count' => $cacheRecordCount, 'n' => $cacheRecordCount]
-                ),
-                'value' => $site->id,
+            $sites[$site->id] = [
+                'name' => $site->name,
+                'path' => $path,
+                'count' => $count,
             ];
         }
 
+        foreach ($sites as $id => $site) {
+            // Count subpaths so we can get an accurate count for this site's path only
+            $subpathCount = 0;
+
+            foreach ($sites as $otherId => $otherSite) {
+                if ($otherId != $id && strpos($otherSite['path'], $site['path']) === 0) {
+                    $subpathCount += is_dir($otherSite['path']) ? count(FileHelper::findFiles($otherSite['path'])) : 0;
+                }
+            }
+
+            $sites[$id]['count'] = $sites[$id]['count'] - $subpathCount;
+        }
+
         return Craft::$app->getView()->renderTemplate('blitz/_utility', [
-            'options' => $options,
+            'sites' => $sites,
         ]);
     }
 }